refactor:#200 redirect: true일 때로 복귀
